refactor :  Result 클래스 삭제 및 Location클래스 추가
